Bug Fix: SettingsViewController instances are not deallocated
"destroy" operation was not called in the current Tchap implementation, but we don't need it anymore.
Indeed removing all the implicit retains of "self" fixes the issue.
CAUTION: we retain explicitly "self" for the 3 following operations: saving, pwd reset and msisdn binding, in order to let them running when the user leaves the settings screen.

Note: we clean here MediaPickerViewController too, by removing all the implicit retains of "self".

"destroy" operation is not anymore required to release the instances of the following class: SettingsViewController, MediaPickerViewController, WebViewViewController, AlbumContentViewController.