Bug Fix - The app may be stuck on the device verification screen
This happens when several check of the pending room key requests are triggered simultaneously.
For example when we resume the app after added a new device.
This multiple requests involve an issue on the dialog handling (The screen is displayed, but the model lost its delegate)

We fix this here by preventing the app from checking several times the pending requests