Bug Fix: [Room Preview] unexpected failure on join: "Not possible to join an empty room"
#346

When the link is based on the room alias, we convert the room alias into a room id to check alias validity.
Then we have to keep the alias to promote join by room alias instead of room id when an alias is available, in order to handle federated room.